% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{box_pack}
\alias{box_pack}
\title{Simulate box packing chain.}
\usage{
box_pack(n, p)
}
\arguments{
\item{n}{number of steps to simulate}

\item{p}{probabilities for item weights}
}
\description{
Simulate box packing chain.
}

