\name{BoxPack-package}
\alias{BoxPack-package}
\alias{BoxPack}
\docType{package}
\title{
\packageTitle{BoxPack}
}
\description{
\packageDescription{BoxPack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BoxPack}
\packageIndices{BoxPack}
~~ An overview of how to use the package, including ~~
~~ the most important functions ~~
}
\author{
\packageAuthor{BoxPack}

Maintainer: \packageMaintainer{BoxPack}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per ~~
~~ line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
