% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pack_boxes}
\alias{pack_boxes}
\title{Simulate box-packing.}
\usage{
pack_boxes(n, p)
}
\arguments{
\item{n}{number of steps to simulate}

\item{p}{vector of probabilities for item weights 1, ..., w}
}
\description{
Simulate box-packing.
}

